Ranged Everything
Ranged Attacks for Enemys and Player
Reworked enemy stats handling to be spesific to the attack type (yes we could have some general enemy stats script and use iheritance and sure that would be good practice for oop but it was easier to to this rework and the effect is the same). This rework breaks nothing in this version and shouldn't affect anything you guys are working on.